# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Visitor API",
        "description": "API Documentation for Visitor Management",
        "contact": {
            "email": "support@mysite.com"
        },
        "license": {
            "name": "Proprietary",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/courses": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create a new course",
                "operationId": "d008cbab8faba4c985410946c75f52fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "starting_price",
                                    "level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Introduction to Programming"
                                    },
                                    "image": {
                                        "description": "Image file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "discounted_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "starting_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "final_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "Pemula",
                                            "Menengah",
                                            "Ahli"
                                        ],
                                        "example": "Pemula"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course created successfully"
                                        },
                                        "course": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Introduction to Programming"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "courses/image.jpg"
                                                },
                                                "starting_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "discounted_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "final_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "level": {
                                                    "type": "string",
                                                    "example": "Pemula"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-10-01T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-10-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/courses/{id}": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update an existing course",
                "operationId": "8f31e3b0cbb50da53fa42e88660fe0f3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "image",
                                    "starting_price",
                                    "level"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Advanced PHP Course"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "discounted_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "starting_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "final_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "level": {
                                        "type": "string",
                                        "enum": [
                                            "Pemula",
                                            "Menengah",
                                            "Ahli"
                                        ],
                                        "example": "Menengah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course updated successfully"
                                        },
                                        "course": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Advanced PHP Course"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "courses/advanced-php-course.jpg"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "advanced-php-course"
                                                },
                                                "starting_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100
                                                },
                                                "discounted_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "final_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50
                                                },
                                                "level": {
                                                    "type": "string",
                                                    "example": "Menengah"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete a course",
                "description": "Deletes a course by its ID",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Course not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}